requirements.txt
beautifulsoup4==4.12.3
homeassistant==2024.3.3
python-dotenv==1.0.1
Requests==2.32.3
voluptuous==0.13.1
